fire-detection-api using flask/requirements.txt
Flask==2.2.5
tensorflow-cpu==2.18.0
numpy
pillow
joblib
werkzeug
gunicorn
